implémentation DAO pour le trajet

+++ src/main/kotlin/com/projet/covoiturage/DAO/TrajetDAOImplMemoire.kt
@@ -1,22 +1,34 @@
 package com.projet.covoiturage.DAO
 
-import com.projet.covoiturage.Model.Reservation
+
 import com.projet.covoiturage.Model.Trajet
-import com.projet.covoiturage.Repository.ReservationRepo
+
 import com.projet.covoiturage.Repository.TrajetRepo
-import org.springframework.jdbc.core.JdbcTemplate
+
 import org.springframework.stereotype.Repository
 
 @Repository
-class TrajetDAOImplMemoire (val db: JdbcTemplate, val repo: TrajetRepo) : TrajetDAO {
+class TrajetDAOImplMemoire ( val repo: TrajetRepo) : TrajetDAO {
+    override fun chercherTrajetsParutilisateur(id: Int): List<Trajet?>?  = repo.findAllTrajetsByUser(id)
+
+    override fun chercherTrajetParUtilisateur(idTrajet: Int, idUtilisateur: Int): Trajet? =
+    repo.findTrajetByUser(idTrajet, idUtilisateur)
+
     override fun chercherTous(): List<Trajet> = repo.findAll();
-    override fun chercherParId(id: Int): Trajet? = repo.findById(id).get()
+    override fun chercherParId(id: Int): Trajet? {
+        if (repo.findById(id).isPresent)
+            return repo.findById(id).get()
+        return null
+    }
 
     override fun ajouter(t: Trajet): Trajet? = repo.save(t)
 
     override fun supprimer(id: Int): Boolean {
-        repo.deleteById(id)
-        return !repo.findById(id).isPresent
+        if (repo.findById(id).isPresent) {
+            repo.deleteById(id)
+            return true
+        } else
+            return false
     }
 
 }
